On branch holiday  Changes to be committed: 	modified:   physics.py 	modified:   server.py
Added system to allow units of the same player moving through each other without breaking their formations as much. It is entirely handled by making modifications to the collision parameters and unit spacing. This is woefully inefficient in achieving it's goal. It seems some local pathfinding is required instead. It is also still a bit buggy. This is only for testing purposes.

- filling the grid of the map now detects if soldiers have allies or enemies close by
- if allies are close (and moving closer, not away) but no enemies they will increase their spacing, slow down and limit their knokback
- when they are away from their allies they reduce their spacing again