Update Pathfinding:
This update adds rudamentary pathfinding for the units. Waypoints can be placed in the map editor. When pressing P or when saving line of sight is evaluated between any pair of waypoints and the Floyd-Warshall algorithm used to determine the shortest paths between any 2 waypoints. This information is stored in the map file and loaded into the game. When giving orders to a unit and there is no line of sight between the positions of two consecutive orders the game will look at all waypoints with line of sight to the first order and all waypoints with line of sight to the second order and determine the fastest possible route.
CAUTION: It is intended that there is a possible path between any two waypoints. Saving with waypoints that are not indirectly connected due to lack of line of sight will crash the map editor.

- Added new type of map object: Waypoint. Has no collisions with soldiers.
- Waypoints and pathfinding-tables are now saved and loaded from map-files.
- Slightly modified how obsolete future orders are deleted when an alternative set of orders is sent to a unit to accomodate pathfinding.
- Added visualization of waypoints and their connections to the map editor.
- Added waypoints to the testmap.
- A new map with a maze was added to test the pathfinding.
- Added new unit Lone Rider. It is a single rider from the Cavalry unit. He is small enough to traverse the new maze map.

Technical:
- Changed detection of line-line intersection to use Bezier parameters because the previous version couldn't handle vertical lines.
- Fixed units other than Infantry not being able to be added to the game when using Arrowkey-Up for cycling.
- Added function Reposition to Circle so that changes to pos are carried over into the Point parent-class.
- Moved functions for line-circle collision and line-rectangle collision into extra_math.

Cosmetic:
- Changed help in the map editor to be a toggle.
- Old location of objects that are being moved are now displayed in dark green in the map editor.
- Added visualization of rectangle diagonals to debug view.